+++ rhel/internal/pulp/manifest.go
