+++ rhel/internal/pulp/manifest_test.go
